+++ cygnss.c
@@ -71,6 +71,7 @@ void readL1data(char L1dataFilename[], int sampleIndex, int ddm_index, struct CY
     l1data->noise_figure = readnc_float_2d(ncid, "lna_noise_figure", sampleIndex, ddm_index);
     l1data->inc_angle = readnc_float_2d(ncid, "sp_inc_angle", sampleIndex, ddm_index);
     l1data->fresnel_coeff2 = readnc_float_2d(ncid, "fresnel_coeff", sampleIndex, ddm_index);
+    l1data->sp_rx_gain = readnc_float_2d(ncid, "sp_rx_gain", sampleIndex, ddm_index);
 
     if (l1data->ddm_ant == 2){
         l1data->ant_temperature_cels = readnc_float_1d(ncid, "lna_temp_nadir_starboard", sampleIndex);