+++ debug.c
@@ -37,9 +37,18 @@ void printfGeometry(geometryData orbitGeometry){
     printf("%f   ",orbitGeometry.g[0].sx_pos_ecef[1]);
     printf("%f   ",orbitGeometry.g[0].sx_pos_ecef[2]);
     printf("\n");
-    //printf("sx_vel = %f   ",orbitGeometry.g[0].sx_vel_ecef[0]);
-    //printf("%f   ",orbitGeometry.g[0].sx_vel_ecef[1]);
-    //printf("%f   ",orbitGeometry.g[0].sx_vel_ecef[2]);
+    printf("rx_vel = %f   ",orbitGeometry.g[0].rx_vel_ecef[0]);
+    printf("%f   ",orbitGeometry.g[0].rx_vel_ecef[1]);
+    printf("%f   ",orbitGeometry.g[0].rx_vel_ecef[2]);
+    printf("\n");
+    printf("tx_vel = %f   ",orbitGeometry.g[0].tx_vel_ecef[0]);
+    printf("%f   ",orbitGeometry.g[0].tx_vel_ecef[1]);
+    printf("%f   ",orbitGeometry.g[0].tx_vel_ecef[2]);
+    printf("\n");
+    printf("sx_vel = %f   ",orbitGeometry.g[0].sx_vel_ecef[0]);
+    printf("%f   ",orbitGeometry.g[0].sx_vel_ecef[1]);
+    printf("%f   ",orbitGeometry.g[0].sx_vel_ecef[2]);
+    printf("\n");
     //printf("\n");
 }