+++ gnssr.h
@@ -96,6 +96,8 @@ typedef struct {
     double SPEC_TO_RX_ORB_FRAME[9], SPEC_TO_TX_ORB_FRAME[9];
     double SPEC_TO_RX_ORB_FRAME_AARON[9], SPEC_TO_TX_ORB_FRAME_AARON[9];
 
+    double ECEF_TO_RX_ORB_FRAME[9], ECEF_TO_TX_ORB_FRAME[9];
+
     // various angles between things
     double angleSxFromRx_rad[2], angleSxFromTx_rad[2];
     double angleSxFromRxAaron_rad[2], angleSxFromTxAaron_rad[2];
@@ -129,11 +131,15 @@ typedef struct {
 
 void geom_printToLog(FILE *outputPtr, int index, orbitGeometryStruct *g); // keep it as it may be useful
 
-void getECEF2SpecularFrameXfrm( double rx_pos_ecef[3], double tx_pos_ecef[3], double sx_pos_ecef[3], double M[9]);
 void geom_calculateSecondaryGeometry( orbitGeometryStruct *g );
 void geom_initialize(geometryData *gd, struct Geometry geom);
 void geom_getRelativeAngleInFrame( double origin[3], double pos[3] , double M[9], double angles_rad[2] );
 orbitGeometryStruct *geom_getOrbitData(geometryData *gd, int geomIdx );
+
+void getECEF2SpecularFrameXfrm( double rx_pos_ecef[3], double tx_pos_ecef[3], double sx_pos_ecef[3], double xfrmMatrix[9]);
+void getSpecularFrameToOrbitFrameXfrm( double sat_pos[3], double sat_vel[3], double xfrmMatrix[9] );
+void getSpecularFrameToAaronOrbitFrameXfrm( double sat_pos[3], double sat_vel[3], double xfrmMatrix[9] );
+void getECEF2OrbitFrameXfrm( double sat_pos_ecef[3], double sat_vel_ecef[3], double xfrmMatrix[9] ); // by Feixiong
 
 //******************************************************************************/
 // specular.c