+++ ChangeLog.md
@@ -1,3 +1,10 @@
+# 2020-08-05
+- TxG = 1 in surface.c
+- use Tx_eirp_watt instead of Tx_Power_dB in powerParm structure
+- add sp_rx_gain in cygnss.c 
+- correct the transformation from SPEC to Orbit (big change)
+- index fixed in wind_interpolate (same as Theis)
+
 # 2020-07-22
 - use CYGNSS reflection coefficient (cyg_R2)
 - small update of CYGNSS GMF